requirements.txt
pyTelegramBotAPI==4.3.0
python-dotenv==0.19.2
